Move the image type tests to own test file
